# server/OpenAPI.yaml
openapi: 3.0.0
info:
  title: Plant.ly API
  version: v1
  description: API para monitoramento de plantas.
servers:
  - url: http://localhost:3030
    description: Servidor local para a API

paths:

  #### Usuários ####
  /v1/users:
    post:
      summary: Cadastrar um novo usuário
      description: Cria um novo usuário e retorna o ID e os dados cadastrados
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Nome do usuário
                  example: João da Silva
                email:
                  type: string
                  description: Email do usuário
                  example: joao@email.com
                password:
                  type: string
                  description: Senha do usuário
                  format: password
                  example: SenhaForte123!
                plants:
                  type: array
                  description: Lista de plantas associadas ao usuário
                  items:
                    type: string
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Usuário criado com sucesso!
                  id:
                    type: string
                    example: 64fd87ef7345b
                  user:
                    type: object
                    properties:
                      name:
                        type: string
                        example: João da Silva
                      email:
                        type: string
                        example: joao@email.com
                      plants:
                        type: array
                        items:
                          type: string
        '500':
          description: Erro ao criar o usuário

  /v1/users/{id}:
    get:
      summary: Buscar usuário por ID
      description: Retorna os detalhes de um usuário com base no ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: 64fd87ef7345b
      responses:
        '200':
          description: Dados do usuário encontrados
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: 64fd87ef7345b
                  name:
                    type: string
                    example: João da Silva
                  email:
                    type: string
                    example: joao@email.com
                  plants:
                    type: array
                    items:
                      type: string
        '404':
          description: Usuário não encontrado
        '500':
          description: Erro ao buscar o usuário

  #### Autenticação ####
  /v1/auth/login:
    post:
      summary: Login de usuário
      description: Faz login e autentica o usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: Email do usuário
                  example: joao@email.com
                password:
                  type: string
                  description: Senha do usuário
                  format: password
                  example: SenhaForte123!
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login realizado com sucesso!
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 64fd87ef7345b
                      name:
                        type: string
                        example: João da Silva
                      email:
                        type: string
                        example: joao@email.com
        '404':
          description: Usuário não encontrado
        '401':
          description: Senha incorreta
        '500':
          description: Erro no servidor

  /v1/auth/reset_password:
    patch:
      summary: Alterar senha
      description: Atualiza a senha do usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: joao@email.com
                current_password:
                  type: string
                  example: SenhaAntiga123!
                new_password:
                  type: string
                  example: NovaSenha123!
                password_validation:
                  type: string
                  example: NovaSenha123!
      responses:
        '200':
          description: Senha alterada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Senha alterada com sucesso!
        '404':
          description: Usuário não encontrado
        '400':
          description: Senhas não coincidem
        '500':
          description: Erro no servidor

  #### Plantas ####
  /v1/plants:
    post:
      summary: Cadastrar uma nova planta
      description: Associa a planta a um usuário e uma espécie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  description: ID do usuário
                  example: 64fd87ef7345b
                name:
                  type: string
                  description: Nome da planta
                  example: Cebolinha
                type:
                  type: string
                  description: Tipo da planta (Espécie)
                  example: Hortaliça
      responses:
        '201':
          description: Planta cadastrada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  plant_id:
                    type: string
                    example: 758fdc4eb72
                  message:
                    type: string
                    example: Planta cadastrada com sucesso e associada ao usuário e à espécie
        '404':
          description: Usuário ou espécie não encontrada
        '500':
          description: Erro ao cadastrar a planta

  /v1/plants/{id}:
    get:
      summary: Buscar detalhes de uma planta
      description: Retorna os detalhes de uma planta com base no ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: 758fdc4eb72
      responses:
        '200':
          description: Dados da planta encontrados
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: 758fdc4eb72
                  name:
                    type: string
                    example: Cebolinha
                  type:
                    type: string
                    example: Hortaliça
        '404':
          description: Planta não encontrada
        '500':
          description: Erro ao buscar a planta

    delete:
      summary: Deletar uma planta
      description: Exclui uma planta com base no ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: 758fdc4eb72
      responses:
        '200':
          description: Planta deletada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Planta deletada com sucesso
        '404':
          description: Planta não encontrada
        '500':
          description: Erro ao deletar a planta

  #### Espécies ####
  /v1/vegetables:
    post:
      summary: Cadastrar uma nova espécie
      description: Cadastra uma nova espécie de planta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Hortelã
                scientific_name:
                  type: string
                  example: Mentha spicata
                watering_tips:
                  type: string
                  example: Regar 2 vezes por semana
                sunlight_tips:
                  type: string
                  example: Luz solar direta
                soil_tips:
                  type: string
                  example: Solo fértil e bem drenado
                temperature_min:
                  type: number
                  example: 10
                temperature_max:
                  type: number
                  example: 30
                growth_time:
                  type: string
                  example: 3 meses
      responses:
        '201':
          description: Espécie cadastrada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Espécie cadastrada com sucesso
                  species:
                    type: object
                    properties:
                      name:
                        type: string
                        example: Hortelã
                      scientific_name:
                        type: string
                        example: Mentha spicata
                      watering_tips:
                        type: string
                        example: Regar 2 vezes por semana
        '500':
          description: Erro ao cadastrar a espécie

  /v1/vegetables/{id}:
    get:
      summary: Buscar dados de uma espécie
      description: Retorna os detalhes de uma espécie
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: 648a4f2eae23
      responses:
        '200':
          description: Dados da espécie encontrados
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    example: Hortelã
                  scientific_name:
                    type: string
                    example: Mentha spicata
                  watering_tips:
                    type: string
                    example: Regar 2 vezes por semana
                  sunlight_tips:
                    type: string
                    example: Luz solar direta
                  soil_tips:
                    type: string
                    example: Solo fértil e bem drenado
        '404':
          description: Espécie não encontrada
        '500':
          description: Erro ao buscar a espécie

    delete:
      summary: Deletar uma espécie
      description: Exclui uma espécie pelo ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: 648a4f2eae23
      responses:
        '200':
          description: Espécie deletada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Espécie deletada com sucesso
        '404':
          description: Espécie não encontrada
        '500':
          description: Erro ao deletar a espécie

  #### Sensores ####
  /v1/sensors:
    get:
      summary: Buscar dados dos sensores
      description: Retorna dados dos sensores com filtros por planta, data, temperatura e umidade
      parameters:
        - name: plant_id
          in: query
          required: true
          schema:
            type: string
            example: 64fd87ef7345b
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
            example: 2024-10-01T00:00:00Z
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
            example: 2024-10-10T23:59:59Z
        - name: temperature_min
          in: query
          schema:
            type: number
            example: 20.0
        - name: temperature_max
          in: query
          schema:
            type: number
            example: 30.0
        - name: humidity_min
          in: query
          schema:
            type: number
            example: 40.0
        - name: humidity_max
          in: query
          schema:
            type: number
            example: 70.0
        - name: soil_moisture_status
          in: query
          schema:
            type: string
            example: Úmido
      responses:
        '200':
          description: Lista de dados dos sensores filtrados
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    plant:
                      type: string
                      example: Cebolinha
                    temperature_air:
                      type: number
                      example: 26.7
                    humidity_air:
                      type: number
                      example: 54.4
                    soil_moisture:
                      type: string
                      example: Úmido
                    timestamp:
                      type: string
                      format: date-time
                      example: 2024-10-10T14:46:14Z
                    week:
                      type: integer
                      example: 1
        '500':
          description: Erro ao buscar dados dos sensores